+++ addons/coupon/data/coupon_email_data.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" ?>
 <odoo>
    <data noupdate="1">
-      <record id="sale_coupon.mail_template_sale_coupon" model="mail.template">
+      <record id="coupon.mail_template_coupon" model="mail.template">
          <field name="name">Coupon: Send by Email</field>
-         <field name="model_id" ref="sale_coupon.model_sale_coupon"/>
+         <field name="model_id" ref="coupon.model_coupon_coupon"/>
          <field name="subject">Your reward coupon from ${object.program_id.company_id.name} </field>
          <field name="email_from">${object.program_id.company_id.email | safe}</field>
          <field name="partner_to">${object.partner_id.id}</field>
