+++ addons/coupon/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import coupon_rules
+from . import coupon_reward
+from . import coupon_program
+from . import coupon_coupon
