+++ addons/coupon/models/coupon_coupon.py
@@ -1,0 +1,73 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+import random
+from dateutil.relativedelta import relativedelta
+
+from odoo import api, fields, models, _
+
+
+class CouponCoupon(models.Model):
+    _name = 'coupon.coupon'
+    _description = "Coupon"
+    _rec_name = 'code'
+
+    @api.model
+    def _generate_code(self):
+        """Generate a 20 char long pseudo-random string of digits for barcode
+        generation.
+
+        A decimal serialisation is longer than a hexadecimal one *but* it
+        generates a more compact barcode (Code128C rather than Code128A).
+
+        Generate 8 bytes (64 bits) barcodes as 16 bytes barcodes are not
+        compatible with all scanners.
+         """
+        return str(random.getrandbits(64))
+
+    code = fields.Char(default=_generate_code, required=True, readonly=True)
+    expiration_date = fields.Date('Expiration Date', compute='_compute_expiration_date')
+    state = fields.Selection([
+        ('reserved', 'Reserved'),
+        ('new', 'Valid'),
+        ('used', 'Consumed'),
+        ('expired', 'Expired')
+        ], required=True, default='new')
+    partner_id = fields.Many2one('res.partner', "For Customer")
+    program_id = fields.Many2one('coupon.program', "Program")
+    discount_line_product_id = fields.Many2one('product.product', related='program_id.discount_line_product_id', readonly=False,
+        help='Product used in the sales order to apply the discount.')
+
+    _sql_constraints = [
+        ('unique_coupon_code', 'unique(code)', 'The coupon code must be unique!'),
+    ]
+
+    def _compute_expiration_date(self):
+        for coupon in self.filtered(lambda x: x.program_id.validity_duration > 0):
+            coupon.expiration_date = (coupon.create_date + relativedelta(days=coupon.program_id.validity_duration)).date()
+
+    def action_coupon_sent(self):
+        """ Open a window to compose an email, with the edi invoice template
+            message loaded by default
+        """
+        self.ensure_one()
+        template = self.env.ref('coupon.mail_template_coupon', False)
+        compose_form = self.env.ref('mail.email_compose_message_wizard_form', False)
+        ctx = dict(
+            default_model='coupon.coupon',
+            default_res_id=self.id,
+            default_use_template=bool(template),
+            default_template_id=template.id,
+            default_composition_mode='comment',
+            custom_layout='mail.mail_notification_light',
+        )
+        return {
+            'name': _('Compose Email'),
+            'type': 'ir.actions.act_window',
+            'view_mode': 'form',
+            'res_model': 'mail.compose.message',
+            'views': [(compose_form.id, 'form')],
+            'view_id': compose_form.id,
+            'target': 'new',
+            'context': ctx,
+        }
