+++ addons/coupon/report/coupon_report.py
@@ -5,15 +5,15 @@ from odoo import api, models
 
 
 class CouponReport(models.AbstractModel):
-    _name = 'report.sale_coupon.report_coupon'
-    _description = 'Sales Coupon Report'
+    _name = 'report.coupon.report_coupon'
+    _description = 'Coupon Report'
 
     @api.model
     def _get_report_values(self, docids, data=None):
-        docs = self.env['sale.coupon'].browse(docids)
+        docs = self.env['coupon.coupon'].browse(docids)
         return {
             'doc_ids': docs.ids,
-            'doc_model': 'sale.coupon',
+            'doc_model': 'coupon.coupon',
             'data': data,
             'docs': docs,
         }
