+++ addons/coupon/report/coupon_report.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <report
+            id="report_coupon_code"
+            model="coupon.coupon"
+            name="coupon.report_coupon_i18n"
+            file="coupon.report_coupon_i18n"
+            report_type="qweb-pdf"
+            string="Coupon Code"/>
+    </data>
+</odoo>
