+++ addons/coupon/security/ir.model.access.csv
@@ -1,0 +1,9 @@
+id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
+access_coupon_coupon_user,user coupon,model_coupon_coupon,sales_team.group_sale_salesman,1,0,0,0
+access_coupon_coupon_manager,manager coupon,model_coupon_coupon,sales_team.group_sale_manager,1,1,1,0
+access_coupon_program_user,user program,model_coupon_program,sales_team.group_sale_salesman,1,0,0,0
+access_coupon_program_manager,manager program,model_coupon_program,sales_team.group_sale_manager,1,1,1,1
+access_coupon_reward_user,user reward,model_coupon_reward,sales_team.group_sale_salesman,1,0,0,0
+access_coupon_reward_manager,manager reward,model_coupon_reward,sales_team.group_sale_manager,1,1,1,0
+access_coupon_rule_user,user rule,model_coupon_rule,sales_team.group_sale_salesman,1,0,0,0
+access_coupon_rule_manager,manager rule,model_coupon_rule,sales_team.group_sale_manager,1,1,1,0
