+++ addons/coupon/views/coupon_program_promo_views.xml
@@ -1,0 +1,106 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <!-- Promotion Program -->
+
+    <record id="coupon_program_view_promo_program_form" model="ir.ui.view">
+        <field name="name">coupon.coupon.promotion.program.form</field>
+        <field name="model">coupon.program</field>
+        <field name="inherit_id" ref="coupon_program_view_form_common"/>
+        <field name="mode">primary</field>
+        <field name="arch" type="xml">
+            <xpath expr="//div[@name='title']" position="inside">
+                <h1><field name="name" class="oe_title" placeholder="Promotion Program Name..." height="20px"/></h1>
+            </xpath>
+            <xpath expr="//widget[@name='web_ribbon']" position="before">
+                <div class="oe_button_box" name="button_box">
+                    <button class="oe_stat_button" type="action" icon="fa-ticket" name="%(coupon_coupon_action)d" attrs="{'invisible': [('promo_applicability', '=', 'on_current_order')]}">
+                        <field name="coupon_count" string="Coupons" widget="statinfo"/>
+                    </button>
+                </div>
+            </xpath>
+            <xpath expr="//group[@name='reward']" position="before">
+                <field name="sequence" invisible="1"/>
+            </xpath>
+            <xpath expr="//field[@name='program_type']" position="after">
+                <field name="rule_partners_domain" placeholder="Select customer" widget="domain" options="{'model': 'res.partner', 'in_dialog': true}"/>
+            </xpath>
+            <xpath expr="//div[@name='rule_minimum_amount']" position="after">
+                <field name="promo_code_usage" widget="radio"/>
+                <field name="promo_code" attrs="{'required': [('promo_code_usage', '=', 'code_needed')], 'invisible': [('promo_code_usage', '=', 'no_code_needed')]}"/>
+            </xpath>
+            <xpath expr="//group[@name='validity']" position="inside">
+                <label string="Apply on First" for="maximum_use_number" class="oe_inline"/>
+                <div>
+                    <field name="maximum_use_number" class="oe_inline"/>
+                    <span> Orders</span>
+                    <span class="oe_grey"> if 0, infinite use</span>
+                </div>
+                <field name="rule_date_from" class="oe_inline"/>
+                <field name="rule_date_to" class="oe_inline"/>
+            </xpath>
+            <xpath expr="//group[@name='reward']" position="before">
+                <group>
+                    <field name="promo_applicability" widget="radio"/>
+                </group>
+                <group>
+                    <span class="oe_grey">
+                        <b>Apply on Current Order -</b> Reward will be applied on current order.<br/>
+                        <b>Apply on Next Order -</b> Generate a coupon for a next order.
+                    </span>
+                </group>
+            </xpath>
+        </field>
+    </record>
+
+    <record id="coupon_program_view_promo_program_tree" model="ir.ui.view">
+        <field name="name">coupon.coupon.promotion.program.tree</field>
+        <field name="model">coupon.program</field>
+        <field name="arch" type="xml">
+            <tree>
+                <field name="sequence" widget="handle"/>
+                <field name="name"/>
+                <field name="promo_code_usage"/>
+                <field name="active"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="coupon_program_view_promo_program_search" model="ir.ui.view">
+        <field name="name">coupon.coupon.promotion.program.search</field>
+        <field name="model">coupon.program</field>
+        <field name="arch" type="xml">
+            <search>
+                <field name="name"/>
+                <filter string="Active" name="active" domain="[('active', '=', True)]" help="Opened Programs"/>
+                <filter string="Closed" name="closed" domain="[('active', '=', False)]" help="Closed Programs"/>
+            </search>
+        </field>
+    </record>
+
+    <record id="coupon_program_action_promo_program" model="ir.actions.act_window">
+        <field name="name">Promotion Programs</field>
+        <field name="res_model">coupon.program</field>
+        <field name="view_mode">tree,kanban,form</field>
+        <field name="view_ids" eval="[
+            (5, 0, 0),
+            (0, 0, {'view_mode': 'tree'}),
+            (0, 0, {'view_mode': 'form', 'view_id': ref('coupon_program_view_promo_program_form')})]"/>
+        <field name="search_view_id" ref="coupon_program_view_promo_program_search"/>
+        <field name="domain">[('program_type', '=', 'promotion_program')]</field>
+        <field name="context">{
+            'default_program_type': 'promotion_program',
+            'default_promo_code_usage':
+            'no_code_needed',
+            'search_default_opened': 1
+            }</field>
+        <field name="help" type="html">
+            <p class="o_view_nocontent_smiling_face">
+                Create a new promotion program
+            </p><p>
+                Build up promotion programs to attract more customers with discounts, free products, free delivery, etc.
+                You can share promotion codes or grant the promotions automatically if some conditions are met.
+             </p>
+        </field>
+    </record>
+</odoo>
