+++ addons/coupon/wizard/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import sale_coupon_report
+from . import coupon_generate
