+++ addons/coupon/wizard/coupon_generate.py
@@ -4,9 +4,9 @@ from odoo import api, fields, models
 from odoo.tools.safe_eval import safe_eval
 
 
-class SaleCouponGenerate(models.TransientModel):
-    _name = 'sale.coupon.generate'
-    _description = 'Generate Sales Coupon'
+class CouponGenerate(models.TransientModel):
+    _name = 'coupon.coupon.generate'
+    _description = 'Generate Coupon'
 
     nbr_coupons = fields.Integer(string="Number of Coupons", help="Number of coupons", default=1)
     generation_type = fields.Selection([
@@ -18,19 +18,19 @@ class SaleCouponGenerate(models.TransientModel):
     def generate_coupon(self):
         """Generates the number of coupons entered in wizard field nbr_coupons
         """
-        program = self.env['sale.coupon.program'].browse(self.env.context.get('active_id'))
+        program = self.env['coupon.program'].browse(self.env.context.get('active_id'))
 
         vals = {'program_id': program.id}
 
         if self.generation_type == 'nbr_coupon' and self.nbr_coupons > 0:
             for count in range(0, self.nbr_coupons):
-                self.env['sale.coupon'].create(vals)
+                self.env['coupon.coupon'].create(vals)
 
         if self.generation_type == 'nbr_customer' and self.partners_domain:
             for partner in self.env['res.partner'].search(safe_eval(self.partners_domain)):
                 vals.update({'partner_id': partner.id})
-                coupon = self.env['sale.coupon'].create(vals)
+                coupon = self.env['coupon.coupon'].create(vals)
                 subject = '%s, a coupon has been generated for you' % (partner.name)
-                template = self.env.ref('sale_coupon.mail_template_sale_coupon', raise_if_not_found=False)
+                template = self.env.ref('coupon.mail_template_coupon', raise_if_not_found=False)
                 if template:
                     template.send_mail(coupon.id, email_values={'email_to': partner.email, 'email_from': self.env.user.email or '', 'subject': subject,})
