+++ addons/coupon/wizard/coupon_generate_views.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="sale_coupon_generate_view_form" model="ir.ui.view">
-        <field name="name">sale.coupon.generate.form</field>
-        <field name="model">sale.coupon.generate</field>
+    <record id="coupon_coupon_generate_view_form" model="ir.ui.view">
+        <field name="name">coupon.coupon.generate.form</field>
+        <field name="model">coupon.coupon.generate</field>
         <field name="arch" type="xml">
             <form string="Generate Coupons">
                 <group>
@@ -18,11 +18,11 @@
         </field>
     </record>
 
-    <record id="sale_coupon_generate_action" model="ir.actions.act_window">
+    <record id="coupon_generate_action" model="ir.actions.act_window">
         <field name="name">Number of Coupons To Generate</field>
-        <field name="res_model">sale.coupon.generate</field>
+        <field name="res_model">coupon.coupon.generate</field>
         <field name="view_mode">form</field>
         <field name="target">new</field>
-        <field name="view_id" ref="sale_coupon_generate_view_form"/>
+        <field name="view_id" ref="coupon_coupon_generate_view_form"/>
     </record>
 </odoo>
