+++ addons/sale_coupon/models/__init__.py
@@ -2,7 +2,5 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import sale_coupon
-from . import sale_coupon_reward
-from . import sale_coupon_rules
 from . import sale_coupon_program
 from . import sale_order
