--- addons/sale_coupon/report/sale_coupon_report.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <report
-            id="report_coupon_code" 
-            model="sale.coupon"
-            name="sale_coupon.report_coupon_i18n"
-            file="sale_coupon.report_coupon_i18n"
-            report_type="qweb-pdf"
-            string="Coupon Code"/>
-    </data>
-</odoo>
