+++ addons/sale_coupon/views/res_config_settings_views.xml
@@ -9,10 +9,10 @@
             <xpath expr="//div[@id='sale_coupon']" position="after">
                 <div class="content-group">
                     <div class="mt8" attrs="{'invisible': [('module_sale_coupon', '=', False)]}">
-                        <button name="%(sale_coupon.sale_coupon_program_action_promo_program)d" icon="fa-arrow-right" type="action" string="Promotion Programs" class="btn-link"/>
+                        <button name="%(coupon.coupon_program_action_promo_program)d" icon="fa-arrow-right" type="action" string="Promotion Programs" class="btn-link"/>
                     </div>
                     <div class="mt8" attrs="{'invisible': [('module_sale_coupon', '=', False)]}">
-                        <button name="%(sale_coupon.sale_coupon_program_action_coupon_program)d" icon="fa-arrow-right" type="action" string="Coupon Programs" class="btn-link"/>
+                        <button name="%(coupon.coupon_program_action_coupon_program)d" icon="fa-arrow-right" type="action" string="Coupon Programs" class="btn-link"/>
                     </div>
                  </div>
              </xpath>
