+++ addons/sale_coupon/views/sale_coupon_views.xml
@@ -1,48 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="sale_coupon_view_tree" model="ir.ui.view">
-        <field name="name">sale.coupon.tree</field>
-        <field name="model">sale.coupon</field>
+    <record id="coupon_coupon_view_tree" model="ir.ui.view">
+        <field name="name">coupon.coupon.tree</field>
+        <field name="model">coupon.coupon</field>
+        <field name="inherit_id" ref="coupon.coupon_coupon_view_tree"/>
         <field name="arch" type="xml">
-            <tree string="Coupons" create="false" edit="false" delete="false">
-                <field name="code"/>
-                <field name="expiration_date"/>
-                <field name="program_id"/>
-                <field name="partner_id"/>
+            <xpath expr="//field[@name='partner_id']" position="after">
                 <field name="order_id"/>
-                <field name="state"/>
-            </tree>
+            </xpath>
         </field>
     </record>
 
-    <record id="sale_coupon_action" model="ir.actions.act_window">
-        <field name="name">Coupons</field>
-        <field name="res_model">sale.coupon</field>
-        <field name="view_id" ref="sale_coupon_view_tree"/>
-        <field name="domain">[('program_id', '=', active_id)]</field>
-        <field name="context">{}</field>
-    </record>
-
-
-    <record id="sale_coupon_view_form" model="ir.ui.view">
-        <field name="name">sale.coupon.form</field>
-        <field name="model">sale.coupon</field>
+    <record id="coupon_coupon_view_form" model="ir.ui.view">
+        <field name="name">coupon.coupon.form</field>
+        <field name="model">coupon.coupon</field>
+        <field name="inherit_id" ref="coupon.coupon_coupon_view_form"/>
         <field name="arch" type="xml">
-            <form string="Coupons" create="false" edit="false" delete="false">
-                <header>
-                     <button name="action_coupon_sent" type="object" string="Send by Email" class="oe_highlight"/>
-                    <field name="state" widget="statusbar" statusbar_visible="new,used,expired" context="{'state': state}"/>
-                </header>
-                <sheet>
-                    <group>
-                        <field name="code"/>
-                        <field name="expiration_date"/>
-                        <field name="partner_id"/>
-                        <field name="order_id"/>
-                        <field name="sales_order_id"/>
-                    </group>
-                </sheet>
-            </form>
+            <xpath expr="//field[@name='partner_id']" position="after">
+                <field name="order_id"/>
+                <field name="sales_order_id"/>
+            </xpath>
         </field>
     </record>
